Move ensure IDs handler function
